% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{arrayEval}
\alias{arrayEval}
\title{Evaluate Array at Specified margin}
\usage{
arrayEval(A, mar)
}
\arguments{
\item{A}{array}

\item{mar}{list of vectors of specified margin}
}
\value{
Array compressed my mar
}
\description{
Evaluate Array at Specified margin
}
