% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-duplicate.R
\name{duplicated_rows}
\alias{duplicated_rows}
\title{Duplicated Row Index Location
Fing the index B's rows from matrix A}
\usage{
duplicated_rows(A, B)
}
\arguments{
\item{A}{rows you wish to locate in B}

\item{B}{larger matrix that contains rows from A}
}
\value{
logical of length nrow(B)
}
\description{
Duplicated Row Index Location
Fing the index B's rows from matrix A
}
