% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps-allocation.R
\name{geo_of_mar}
\alias{geo_of_mar}
\title{Check Geography associated with given margin}
\usage{
geo_of_mar(mar)
}
\arguments{
\item{mar}{list of length 6 with the 6th being Geo margin}
}
\value{
Geo level in ('state', 'county', 'tract')
}
\description{
Check Geography associated with given margin
}
