% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps-allocation.R
\name{geoChar2geoMod}
\alias{geoChar2geoMod}
\title{Convert geo character to epsilon modifier}
\usage{
geoChar2geoMod(geoMod, geo)
}
\arguments{
\item{geoMod}{vector of modifier levels from largest to smallest in
geographic size}

\item{geo}{character valued geography}
}
\value{
entry of geoMod associated with geo
}
\description{
Convert geo character to epsilon modifier
}
