% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood.R
\name{nhood}
\alias{nhood}
\title{Neighborhood of specific cell in array}
\usage{
nhood(
  A,
  idx,
  ownRent_vars = c("own", "rnt"),
  sex_vars = c("mal", "fem"),
  cenRace_vars = c("wh", "bl", "as", "aian", "pac", "oth", "twp"),
  age_vars = c("0-17", "18-62", "62p"),
  hispDes_vars = c("hisp", "nonhisp"),
  minCount = -10^5,
  maxCount = 10^5
)
}
\arguments{
\item{A}{data array}

\item{idx}{specific cell to find its neighborhood}

\item{ownRent_vars}{levels of ownRent variable to return}

\item{sex_vars}{levels of sex variable to return}

\item{cenRace_vars}{levels of CenRace variable to return}

\item{age_vars}{levels of age variable to return}

\item{hispDes_vars}{levels of hisp variable to return}

\item{minCount}{min value to return}

\item{maxCount}{max value to return}
}
\value{
data frame with neighborhood cell of idx
}
\description{
Neighborhood of specific cell in array
}
