% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_lap.R
\name{rand_lap}
\alias{rand_lap}
\title{Random Laplace}
\usage{
rand_lap(n, bpar)
}
\arguments{
\item{n}{number of observations returned}

\item{bpar}{scale parameter of Laplace}
}
\value{
n random laplace variable
}
\description{
Random Laplace
}
\examples{
rand_lap(20, 10)

}
