% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim3way.R
\name{Sim3Way}
\alias{Sim3Way}
\title{Replicate Laplace Noise for 3-way table}
\usage{
Sim3Way(Nrep, intab, bpar, marPack, W = NULL)
}
\arguments{
\item{Nrep}{Number of replication}

\item{intab}{True cell values}

\item{bpar}{Laplace noise parameter}

\item{marPack}{list of lists of vectors with margins to put in model}

\item{W}{(optional) weight vector of same length as d + length(marPack) = total number of predictors.}
}
\value{
A list with the following:
\itemize{
  \item param - mean parameter estimates of Nrep replications
  \item rmse - Root mean square error of param estimates
}
}
\description{
Replicate Laplace Noise for 3-way table
}
