% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim6way.R
\name{Sim6Way}
\alias{Sim6Way}
\title{Replicate Laplace Noise for 6-way table under the exact schema
(hhgq, sex, Cenrace, age, hisp, geo).}
\usage{
Sim6Way(Nrep, intab, eps, marPack, geoMod, queryMod, W = NULL)
}
\arguments{
\item{Nrep}{Number of replication}

\item{intab}{True values - array with 6 dimensions (last is geo)}

\item{eps}{Laplace noise parameter. Privacy Budget}

\item{marPack}{list of lists of vectors with margins to put in model.
Workload of queries}

\item{geoMod}{vector of weights for each geography in marPack}

\item{queryMod}{vector of weights for each query in marPack}

\item{W}{(optional) weight vector of same length as
d + length(marPack) = total number of predictors.}
}
\value{
A list with the following:
\itemize{
  \item param - mean parameter estimates of Nrep replications
  \item rmse - Root mean square error of param estimates
  \item coefEsts - each row is coefficient estimates of a replicate
  \item noise - each column is the noise applied to the true table values
}
}
\description{
Replicate Laplace Noise for 6-way table under the exact schema
(hhgq, sex, Cenrace, age, hisp, geo).
}
